+++ src/index/inverted_index.cpp
@@ -68,7 +68,7 @@ uint64_t inverted_index::idf(term_id t_id) const
     return pdata->inverse_frequency();
 }
 
-uint64_t inverted_index::term_freq(term_id t_id, doc_id d_id) const
+double inverted_index::term_freq(term_id t_id, doc_id d_id) const
 {
     auto pdata = search_primary(t_id);
     return pdata->count(d_id);
@@ -85,18 +85,18 @@ uint64_t inverted_index::total_corpus_terms()
     return _total_corpus_terms;
 }
 
-uint64_t inverted_index::total_num_occurences(term_id t_id) const
+double inverted_index::total_num_occurences(term_id t_id) const
 {
     auto pdata = search_primary(t_id);
 
-    uint64_t sum = 0;
+    double sum = 0;
     for(auto & c: pdata->counts())
         sum += c.second;
 
     return sum;
 }
 
-const std::unordered_map<doc_id, uint64_t>
+const std::unordered_map<doc_id, double>
 inverted_index::counts(term_id t_id) const
 {
     auto pdata = search_primary(t_id);
